[refactor] add Event component and Event folder
