+++ Reactify/ClientApp/src/App.js
@@ -5,7 +5,7 @@ import MusicPlayer from './components/musicPlayer/MusicPlayer';
 import NavBar from './components/navBar/NavBar.js';
 import Footer from './components/footer/Footer.js';
 import BandDetail from './components/bandDetail/BandDetail';
-import EventsList from './components/EventsList';
+import EventsList from './components/Event/EventsList';
 
 import './custom.css'
 import TrackList from "./components/trackList/TrackList";
