+++ Reactify/ClientApp/src/components/Event/EventsList.css
